refactor: rename brand-suffix.ts to brand-suffix-shape.ts
Matches the hyphen-shape.ts naming convention. Updates the import paths in content-script.ts, messages.ts, and the extension/shared tests accordingly.
