+++ browser-extensions/chrome/src/ai-spacing/content-script.ts
@@ -1,7 +1,7 @@
 import type { CandidateLabel, ClassifyCandidatesMessage, ClassifyCandidatesResponse } from './messages';
 import type { AmbiguousShape, SettledCandidate, TextEdit } from './shapes/base';
 import { applyTextEdits } from './shapes/base';
-import { brandSuffix } from './shapes/brand-suffix';
+import { brandSuffix } from './shapes/brand-suffix-shape';
 import { hyphenSign } from './shapes/hyphen-shape';
 
 const pangu = window.pangu;
