+++ browser-extensions/chrome/src/ai-spacing/shapes/brand-suffix-shape.ts
